refactor: PR レビュー指摘に対応
- requireAdmin ミドルウェアを .route() の前に配置し実行順序を保証
- CategoryNameDuplicateChecker を コンストラクタ DI に統一
- requireAdmin に user の null チェックを追加
- app.ts の Auth ルート変更意図をコメントで明記
- DomainConflictError のグローバルハンドラー処理についてコメント追加
- statusLabels/difficultyLabels 等の定数を constants.ts に共通化
- proposal-detail-page.tsx を DetailView/EditForm/RejectDialog に分割
