+++ apps/api/src/application/category/create-category-use-case.ts
@@ -2,7 +2,6 @@ import { Id } from "../../domain/id.js";
 import { Category } from "../../domain/category/category.js";
 import { CategoryName } from "../../domain/category/category-name.js";
 import type { CategoryNameDuplicateChecker } from "../../domain/category/category-name-duplicate-checker.js";
-import type { CategoryQueryService } from "../../domain/category/category-query-service.js";
 import type { CategoryRepository } from "../../domain/category/category-repository.js";
 import type { UnitOfWork } from "../unit-of-work.js";
 
@@ -14,14 +13,13 @@ export class CreateCategoryUseCase {
   constructor(
     private uow: UnitOfWork,
     private categoryRepository: CategoryRepository,
-    private categoryQueryService: CategoryQueryService,
     private duplicateChecker: CategoryNameDuplicateChecker,
   ) {}
 
   async execute(input: CreateCategoryInput): Promise<Category> {
     return this.uow.run(async () => {
       const name = CategoryName.create(input.name);
-      await this.duplicateChecker.ensure(name, this.categoryQueryService);
+      await this.duplicateChecker.ensure(name);
 
       const category = Category.create({
         id: Id.random<Category>(),
